WL#9638: Step 46 ---------------- There is a need to use a slightly larger backup data buffer size for LCPs with small REDO logs. At the same time it is essential to not use a large backup data buffer size when there is no urgency to checkpoint to avoid REDO log overflow.
To handle this we create a 2M backup data buffer size, we still
keep the minimum write size to 256k and the maximum write size
is set to 2M as well. Every time we have completed a scan batch
(when receiving  SCAN_FRAGCONF) we check if we have filled up
such that at least one file can be written. If this is the case
we check that if we have written more than 512k, if this is the
case we only allow the scan to continue if it is an LCP scan and
the REDO alert level is high, at low alert level we allow the
scan to continue filling up to 2M data buffer.

This will provide a balance between ensuring that LCP scans don't
destroy predictable latency and that we don't overflow the
REDO log.