Bug#26876594 MYSQL CRASH:INNODB_API_DISABLE_ROWLOCK=ON&DAEMON_MEMCACHED_R_BATCH_SIZE =1
PROBLEM DESCRIPTION
--------------------

Start the server with innodb_api_disable_rowlock = ON and
daemon_memcached_r_batch_size = 1

Initiate a read from the memcached ( get query)
This read starts a transaction through API and  passes auto_commit
as true since daemon_memcached_r_batch_size is 1.
When innodb_api_disable_rowlock=ON API proceed to create table lock
lock_table_create expects trx_is_autocommit_non_locking
to fail, but this transaction has auto-commit on and
is a nonlocking transaction, so the assert fails.
So it can be concluded that the innodb_api_disable_rowlock=ON is
incompatible with auto-commit.

FIX
---

Pass auto-commit as false when innodb_api_disable_rowlock = ON
and daemon_memcached_r_batch_size = 1

REVIEWED BY: Jimmy Yang<Jimmy.Yang@oracle.com>
 RB: 19403