WL #12074: Volcano iterator executor base [3/15, EXPLAIN]
Add a new format, EXPLAIN FORMAT=TREE, that allows printing out the entire
iterator tree. Every iterator describes itself and lists out which children
it has, if any.

This new format is experimental and subject to change. It does not yet include
any costs or the likes, as we are limited to executor structures, not optimizer
structures.

Note that we currently do not support printing plans for subselects, even if
they are executed by the new executor.

Change-Id: I0f3d078ce6618b968b197d2ad5ad5e9c14ae4ba6