Bug#27442477 ASSERT `THD->GET_HA_DATA...HAS_STATE(XID_STATE::XA_ACTIVE))' AT HANDLER.CC:1396
This is a post-push fix for addressing failing test-scripts.

Problem
-------
Test scripts executed in the same worker thread and just after
`rpl.rpl_xa_with_filters_error_msg` - the original bug-fix patch provided test -
fail to execute both on Jenkins and on PB2.

How to reproduce
----------------

    $ ./mtr --parallel=1 --big-test rpl.rpl_xa_with_filters_error_msg \
      rpl.transactional_ddl_locking

Analysis
--------
The original bug-fix patch analysis concluded that, when using the XA
transaction with replication filters, the server enters an unknown state. That
is exactly the problem here, the next test-script to be executed failed with an
unrelated error.

Fix
---
Reset the filter state, before ending the script.