Bug #27977721: [2/3] Recognize compatible SRS definitions [noclose]
When doing CREATE OR REPLACE SPATIAL REFERENCE SYSTEM and the SRS is in
use by an SRID restriction on a column, check if the statement will
modify any parameters that affect the results of computations. If no
computations are changed, allow the change even if the SRS is in use.

It is allowed to add, but not remove or modify, TOWGS84 clauses from
SRSs that are in use. The rationale for this is that adding a TOWGS84
clause to an SRS that didn't have one before doesn't change any
results. It only opens up for new computations to happen
(transformations). But removing or modifying it will change results.

If the SRS (that is in use) is recognized as WGS 84, allow both adding
and deleting a TOWGS84 clause where all values are 0. No other TOWGS84
clauses can be added to WGS 84 SRSs.

Change-Id: Ic7dc7c274e1dc6a92195f5a254e1b2f67a2da260