BUG#26225783 MYSQL CRASH ON CREATE TABLE (REPRODUCEABLE) -> INNODB: A LONG SEMAPHORE WAIT
PROBLEM
===========
A false upgrade(from SX to X on root page) ends up in a deadlock with
another purge holding S latch on same index tree root. The S latched
thread would be waiting to acquire S latch on non leaf and leaf level
pages that would be X latched by the 1st purge thread and the 1st thread would
be waiting on the 2nd one to release S latch on the root page to proceed
with its upgrade of latch on root page from SX to X.The false upgrade
happens due to a wrong estimation of merge of page of sys table index as
the column would be varchar(0) and it would not consider the empty
bytes while calculating max possible record size.

FIX
===
Added the case when calculating columns whose length isn't defined to
consider those extra bytes when page is being emptied out so that the
estimation isn't wrong at the cursor positioning level.
Also will be raising a separate bug about having a proper length for sys
table columns to avoid such scenarios.

Reviewed by: Debarun Banerjee <debarun.banerjee@oracle.com>
RB: 18851