BUG#27169282 ------------ Local LCP code and distributed LCP code wasn't designed for the case with no fragments in the LDM. Added c_num_fragments that indicates if there are any fragments around when local LCP is started. c_fragments_in_lcp is added to count the number of fragments we have received LCP_FRAG_ORD for. If 0 we don't interact with Backup block in end phase of LCP.
Code was also not designed for the case where MaxParallelCopyInstances was set
to 1. So this fixes both the case when there are no fragments at all in an LDM
thread as well as when an LCP is started before we have received the first
COPY_ACTIVEREQ in the LDM thread.