Bug#27930382: FOREIGN KEY CONSTRAINT VIOLATIONS INCORRECTLY REPORTED
The FK_CHILD / FK_PARENT trigger events, used to check any foreign key
constraints, where executed before the SECONDARY_INDEX triggers which
updates the (unique) secondary indexes.

Thus, even if the row was inserted (or deleted), the row was not
visible throught the unique index is it was not inserted there yet.

A self referrencing foreign key constraint would then not find the
row via the unique indexes, and incorrectly reports a constraint violation
as the (self-) referred row was belived to not exist.

The fix ensure that any FK_CHILD and FK_PARENT triggers are inserted
last in their respective 'triggerLists'. To allow 'last inserts',
the trigger-lists datatype had to be changed from a DLList to a DLFifoList.

Previous bug workaround in testFK.cpp for this bug is removed and
a specific testcase added.