Bug #27847697: UNZIP SDI JSON FILES SO THAT GIT CAN MERGE CHANGES
There are always merge conflicts when merging 2 branches with DD
changes because these SDI JSON files are in 4 zip files. Git cannot
merge the contents of a zip file. This requires manual unzipping,
merging and zipping.

Or it is possible to use the --reccord setting to rewrite the jsons
and zip files, which is not safe since the differences within the
json files are not compared for merge conflicts.

There are 4 json.zip files used in 4 different tests. This patch
unzips these 4 zip files into their own directories under
mysql_test/std_data/dd/sdi/. It also changes these tests to skip
the copy and unzip steps and just compare directly to the appropriate
unzipped json file under std_data/dd/sdi/

The unzipped files take up 3.7MB on disk, which is small compared
to the 320.5 MB mysql-test directory.

Approved by Anna in RB#19535