Bug #28094147: PREPARATORY CHANGES FOR WL #12074 [AlternativeIterator, noclose]
In the case of IN-to-EXISTS rewriting with a pushed condition,
we need to disable said condition, and also all ref accesses that
stem from it, if a NULL value is sought. Instead of rewriting the
plan at runtime, make a new iterator (AlternativeIterator) that can
defer to either a RefIterator or TableScanIterator as needed.

Change-Id: I91801c58e162431544240dfa000ec3f25c180a3a