Bug#24357244: CANNOT CREATE A STORED FUNCTION IF A DATABASE               IS NOT SELECTED
Issue:
------
When default database is not set, a CREATE FUNCTION DDL
that calls another FUNCTION will fail if no explicit
database is specified in the DDL's body.

Solution:
---------
As was done in 5.6, assign the database mentioned in
the new FUNCTION's DDL. If that isn't present, see if
default database is set and use it.