Bug#28040201 - ADDING A NEW COLUMN TO A TABLE HAVING A GENERATED COLUMN CRASHES THE SERVER
This is due to one bugos assertion. The assertion assumes that before one
INSTANT ADD COLUMN, there must be some stored columns in the table already.
However, this is not true if the table has only virtual columns. So the
fix is to simply remove the bogus assertion, since later assertions are
sufficient to check the correction of the remembered instant column numbers.

RB: 19674
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>