Bug#27284699 RENAME USER THROW ERROR IF AUTH_ID IS INHERITED BUT NOT IN              ROLE GRAPH
Description
-----------
The error mentioned in the bug report usually happens when we are trying to
rename the user which already exists. Nevertheless, I noticed the following
error while renaming the role which no longer is granted to any authID.

"ERROR 3532 (HY000): Renaming of a role identifier is forbidden"

Once a vertex is renamed/dropped, we remove all the connecting edges to it.
Let us call this vertex as source and the connecting vertex as target.
A vertex in the role graph is no longer a role if it has no incoming edge.
It means any of target vertex might no longer is a role. Hence we update
the corresponding ACL_USER's is_role flag to false.

Problem here is that the is_role flag is not getting reset to false of the
ACL_USER corresponding to the target vertex unless graph is rebuilt.

Note - The edges direction of the role graph is in reverse of the edges
represented in the mysql.role_edges table. Bug#27915588 is created to fix it.

For instance -

CREATE USER u5;
CREATE ROLE r1,r2,r3;
GRANT r1 TO r2;
GRANT r2 TO r3;
GRANT r3 to u5;

Here is the role graph for the above scenario mentioned in the bug.

u5 --> r3 --> r2 --> r1

DROP USER u5; / REVOKE r3 FROM u5;

After the user u5 is dropped or r3 is revoked from u5 then updated role graph
is as following -

r3 --> r2 --> r1

It is clear that r3 has an adjacent vertex r2 but it does not has any incoming
edge, hence its is_role property must be reset in the corresponding ACL_USER.

Fix:
---
1. To update the is_role flag of ACL USER, we had been checking if the
   target vertex has any adjacent vertex then it is not role. In the boost
   directed graph, adjacent vertices are connected through the outgoing edges.
   On the contrary, we had to look if the target vertex has any incoming edge.
   In order to figure out that the graph has any incoming edge, we needed to
   make the role graph bidirectionalS directed graph. Yes, it increases the
   space overhead but it is needed right now.

2. We must probe the target vertex for incoming edges only after its
   connecting edge from the vertex(to be deleted/renamed) is removed.

3. Improved the current logic to modify the role edges in the table.
   In order to modify the role edges in the table, we used to iterate
   through all the edges in the role graph and check if any of 2 vertices
   of the edge matches with the source vertex to be dropped/renamed.
   Since the role graph is now bidirectionalS directed graph therefore, it is
   possible to just find adjacent vertices connected through the incoming and
   outgoing edges of the source vertex.

Testing :
-------
- Added the scenario reported in this bug in a new test file rename_role.test.
- Also, moved rename role related tests from role1.test to the new test file.

Branch - mysql-8.0-itch
Push Id -  12946637
Timestamp - 2018-04-24 21:37:43

Review : RB#19529Bug#27284699 RENAME USER THROW ERROR IF AUTH_ID IS INHERITED
BUT NOT IN
             ROLE GRAPH

Description
-----------
The error mentioned in the bug report usually happens when we are trying to
rename the user which already exists. Nevertheless, I noticed the following
error while renaming the role which no longer is granted to any authID.

"ERROR 3532 (HY000): Renaming of a role identifier is forbidden"

Once a vertex is renamed/dropped, we remove all the connecting edges to it.
Let us call this vertex as source and the connecting vertex as target.
A vertex in the role graph is no longer a role if it has no incoming edge.
It means any of target vertex might no longer is a role. Hence we update
the corresponding ACL_USER's is_role flag to false.

Problem here is that the is_role flag is not getting reset to false of the
ACL_USER corresponding to the target vertex unless graph is rebuilt.

Note - The edges direction of the role graph is in reverse of the edges
represented in the mysql.role_edges table. Bug#27915588 is created to fix it.

For instance -

CREATE USER u5;
CREATE ROLE r1,r2,r3;
GRANT r1 TO r2;
GRANT r2 TO r3;
GRANT r3 to u5;

Here is the role graph for the above scenario mentioned in the bug.

u5 --> r3 --> r2 --> r1

DROP USER u5; / REVOKE r3 FROM u5;

After the user u5 is dropped or r3 is revoked from u5 then updated role graph
is as following -

r3 --> r2 --> r1

It is clear that r3 has an adjacent vertex r2 but it does not has any incoming
edge, hence its is_role property must be reset in the corresponding ACL_USER.

Fix:
---
1. To update the is_role flag of ACL USER, we had been checking if the
   target vertex has any adjacent vertex then it is not role. In the boost
   directed graph, adjacent vertices are connected through the outgoing edges.
   On the contrary, we had to look if the target vertex has any incoming edge.
   In order to figure out that the graph has any incoming edge, we needed to
   make the role graph bidirectionalS directed graph. Yes, it increases the
   space overhead but it is needed right now.

2. We must probe the target vertex for incoming edges only after its
   connecting edge from the vertex(to be deleted/renamed) is removed.

3. Improved the current logic to modify the role edges in the table.
   In order to modify the role edges in the table, we used to iterate
   through all the edges in the role graph and check if any of 2 vertices
   of the edge matches with the source vertex to be dropped/renamed.
   Since the role graph is now bidirectionalS directed graph therefore, it is
   possible to just find adjacent vertices connected through the incoming and
   outgoing edges of the source vertex.

Testing :
-------
- Added the scenario reported in this bug in a new test file rename_role.test.
- Also, moved rename role related tests from role1.test to the new test file.

Branch - mysql-8.0-itch
Push Id -  12946637
Timestamp - 2018-04-24 21:37:43

Review : RB#19529