WL #12074: Volcano iterator executor base [6/15, materialize derived]
Add support for materializing derived tables.

This also changes DebugString() from returning a single string to returning
a vector of strings; they will be printed out as consecutive children.
This is because we want to report creating the temporary table and scanning
it as two distinct operations.

Note that we this breaks LATERAL support, but it is being fixed in a
separate patch later in the series.

Change-Id: I5605ce721c6360ecc1a103ff796465ac98116ff2