BUG#91011 FIX AUTOTEST-RUN.SH SCRIPT CALL
When autotest-run.sh was modified to use bash instead of standard sh,
autotest-boot.sh got broken due to calling autotest-run.sh with "sh".
Fix autotest-boot.sh to call autotest-run.sh without assuming which
interpreter is used.

NOTE: There are several ways to achieve the above. Using "source" or "."
is not one of them, as that reads the script to the current one and
executes the code. Also, because autotest-run.sh calls exit, that would
prematurely fail autotest-boot.sh. Opting for simply calling the script
and quoting script name to be able to correctly support paths with
spaces.

NOTE 2: "-x" debug option was purposely removed. If more information
needs to be reported that should be improved in autotest-run.sh.