WL #12074: Volcano iterator executor base [7/15, print condition subplans]
When a condition in FilterIterator contains a subquery, print its plan.

This extends the plan printing with children headlines; a RowIterator can now
declare not only that another iterator is a sub-iterator, but also in what
capacity (ie., a headline). This allows us to explain that it comes from
e.g. a condition (in this case).

Headlines are formatted the same way as sub-iterators right now, but that could
easily be changed.

Change-Id: I0b04bc617a1d29e975314b7350d39a624be0b3cb