WL #12074: Volcano iterator executor base [8/15, const tables]
Support queries that include const tables (potentially _only_ const tables),
including the special case of being on the outer join of a const table.
To this end, we add a new type of iterator that returns exactly one row and
then returns EOF (a FakeSingleRowIterator), so that the query sees exactly
one row.

The old executor has a separate query path for when there are no primary
tables, but this seems to be out of need to have somewhere to put the filtering
(which normally would happen in sub_select), not for performance -- at least,
it doesn't seem to give a measurable performance boost if we add a similar path
to the new executor.

Note that this still does not support the special case of no tables at all
(0 primary tables, 0 const tables); that will happen in a later commit.

Change-Id: I73439fec0d80edbf839cca80993e069ff7952e12