Bug#28065244 - MYSQL UPGRADE FROM 8.0.11 TO 8.0.12 MISSING COLS FROM INNODB_TABLES/COLUMNS
The problem is that newly added columns in I_S.innodb_tables and
I_S.innodb_columns can be shown in these two tables themselves.
However, they should also be shown in I_S.COLUMNS, but currently no.

The cause is that server depends on the plugin version change to
decide if the metadata of I_S.COLUMNS should be recreated or not.
Currently, InnoDB maps the server version to this plugin version.
But since the plugin version only accepts the format of X.Y, so the
MYSQL_VERSION_PATCH was always not included into this version.
So an upgrade from 8.0.11 to 8.0.12(or any minor upgrade) would not
recreate the metadata properly, because both 8.0.11 and 8.0.12 would
be remembered as 8.0, then there is no change on the version.

To fix it properly, InnoDB should not use map the server version to
the plugin version by discarding the MYSQL_VERSION_PATCH. For the plugin
version used for InnoDB I_S views, the new format MYSQL_VERSION_MAJOR.X
is used. Here MYSQL_VERSION_MAJOR is always the major version, currently
it's 8. And about the 'X', it doesn't mean the MYSQL_VERSION_MINOR any
more. Instead, it now only means the I_S views structure version,
starting from 1 for now, to distinguish from current 8.0. Every time a
column of InnoDB I_S views gets changed, this X should be increased
by 1, etc.

Please note that the plugin version for 'InnoDB' is not changed in
this patch, it's still the MYSQL_VERSION_MAJOR.MYSQL_VERSION_MINOR.

RB: 19729
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>