BUG#27909771 - [MYSQL 8.0 GA DEBUG BUILD] ASSERTION `(*IDX_EL_IT)->IS_PREFIX() == STATIC_CAST<B
When a table is created using a numeric string of length greater than 65 via
CREATE..SELECT and an index is added on this column,an assertion is triggered
on debug builds under non-strict sql mode (which allows truncated fields).
The assertion happens because the index key length differs
from the character length of the corresponding coulumn object. The char
length exceeds the maximum decimal precision value of decimal type field.
The fix is to use the correct precision value when calculating the key
length.