BUG#27882096 DNS BASED SEEDS RESOLVING TO ITSELF CAUSES GR TO NOT START
Description
=================

When group_replication_group_seeds contains a DNS based entry which resolves
to it's on mysql daemon, group replication will be stuck.

Configure a cluster with the following settings:
group_replication_group_seeds =
node1:33061,192.168.70.3:33061,192.168.70.4:33061
group_replication_local_address = 192.168.70.2:33061

And start it. I won't be able to start.

Analysis
=================

We already had that guard in place, but only if local_address matches seeds,
either by both using name or both using address. If there is a mismatch,
it would not work, as per the example above.

Fix
=================

Always resolve both local_address and seeds to address format before doing the
comparision, and then skip if it matches the local address.