Bug#27903842 - [mysql 8.0 ga release build] handle_fatal_signal(sig=11) in std::__uniq_ptr_impl
While updating column metadata of a view referencing partitioned
table, table handler share(Handler_share) instance is accessed to
read the partition id information using the partition name. But
table is not opened in the storage engine(in MYSQL_OPEN_NO_NEW_TABLE_SE
mode)while updating a referencing view's column metadata. Due to
which pointer to table handler share is not set and accessing it
for partitioned tables resulted in the issue reported.

Since table is not opened in the storage engine and table handler
share instance is not created for it in the MYSQL_OPEN_NO_NEW_TABLE_SE
mode, adding fix to skip read and lock partition bitmap set
for the partition table in MYSQL_OPEN_NO_NEW_TABLE_SE mode.

Change-Id: If929353e9e56007c16ef46baa277db53845b6ea5